+++ apps/web/package.json
@@ -9,6 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@supabase/ssr": "^0.6.1",
+    "@supabase/supabase-js": "^2.53.0",
+    "dotenv": "^17.2.1",
     "lucide-react": "^0.525.0",
     "next": "15.4.1",
     "react": "19.1.0",
