+++ apps/web/src/app/layout.tsx
@@ -1,12 +1,12 @@
-import type { Metadata } from "next";
 import "./globals.css";
+import type { Metadata } from "next";
 
 export const metadata: Metadata = {
   title: "Vibe Three",
   description: "Make 3D Scenes with Vibe Three",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
