+++ apps/web/src/components/home/HeroSection.tsx
@@ -4,10 +4,10 @@ export default function HeroSection() {
     return (
         <>
             <h1 className="text-4xl md:text-6xl font-extrabold text-center text-white mb-4">
-                Vibe in Three-D
+                Vibe in 3D
             </h1>
             <p className="text-lg text-center text-gray-300 mb-6 max-w-2xl">
-                Create 3D scenes and experiences. Describe what you want, bring it to life.
+                Describe what you want, bring it to life.
             </p>
         </>
     );
